services/payment-service/requirements.txt
psycopg2-binary==2.9.9
stripe==7.8.0
paypalrestsdk==1.13.3
boto3==1.34.0
python-dotenv==1.0.0